Add ability to fetch active sprints for a board
This uses the Jira REST API rather than the search interface, so a new
section was needed in the config file.

The hope is that we'll be able to use the results by finding the
sprint with the desired "originBoardId" matching the original query,
and stripping out the part we want (the current sprint number), in
dynamic sheets.